requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
appdirs==1.4.4
asttokens==3.0.1
attrs==25.3.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
bidict==0.23.1
blinker==1.9.0
blis==1.2.1
brotli==1.2.0
build==1.4.0
catalogue==2.0.10
certifi==2026.1.4
cffi==1.17.1
charset-normalizer==3.4.4
chromadb==1.5.5
click==8.2.1
cloudpathlib==0.23.0
colorama==0.4.6
comm==0.2.3
confection==0.1.5
ConfigArgParse==1.7.1
contourpy==1.3.3
crewai==1.6.1
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
emoji==2.15.0
et_xmlfile==2.0.0
executing==2.2.1
faiss-cpu==1.12.0
filelock==3.24.2
filetype==1.2.0
Flask==3.1.0
flask-cors==6.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.7.0
fsspec==2026.2.0
gensim==4.4.0
gevent==25.9.1
geventhttpclient==2.3.8
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-vision==3.12.1
google-genai==1.68.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
installer==0.7.0
instructor==1.14.5
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.3
json5==0.13.0
json_repair==0.25.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kubernetes==35.0.0
langchain==1.2.13
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.21
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langcodes==3.5.0
langdetect==1.0.9
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.4.32
language_data==1.3.0
llvmlite==0.46.0
locust==2.43.3
lxml==6.0.2
marisa-trie==1.3.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.8
matplotlib-inline==0.2.1
mcp==1.26.0
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
mpmath==1.3.0
msal==1.31.1
msgpack==1.1.2
msoffcrypto-tool==6.0.0
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numba==0.63.1
numpy==2.3.3
oauthlib==3.3.1
olefile==0.47
onnx==1.20.1
onnxruntime==1.24.1
openai==2.29.0
opencv-python==4.13.0.92
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
ormsgpack==1.12.2
outcome==1.3.0.post0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
parso==0.8.6
pdf2image==1.17.0
pdfminer.six==20260107
pdfplumber==0.11.7
pi_heif==1.3.0
pikepdf==10.5.1
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
portalocker==2.7.0
posthog==5.4.0
preshed==3.0.10
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==2.22
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.26.5
pyod==2.0.5
pyodbc==5.3.0
pypandoc_binary==1.17
pyparsing==3.3.2
pypdf==6.7.0
PyPDF2==3.0.1
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
PySocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.2
python-engineio==4.13.1
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
python-pptx==1.0.2
python-socketio==5.16.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
qrcode==8.2
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
safetensors==0.6.2
scikit-learn==1.8.0
scipy==1.17.0
selenium==4.38.0
sentence-transformers==5.1.1
setuptools==81.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
sse-starlette==3.2.0
stack-data==0.6.3
starlette==0.52.1
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.12.0
timm==1.0.26
tokenizers==0.22.1
tomli==2.4.0
tomli_w==1.2.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==4.57.1
trio==0.32.0
trio-websocket==0.12.2
typer==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
urllib3==2.5.0
uuid_utils==0.14.1
uv==0.10.3
uvicorn==0.41.0
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.6.0
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.3
wsproto==1.3.2
xlrd==2.0.2
xlsxwriter==3.2.5
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zope.event==6.1
zope.interface==8.2
zstandard==0.25.0
